feat: tratamento de exceções na entidade Music

+++ src/domains/Music/services/MusicService.ts
@@ -1,8 +1,14 @@
 import { Music } from "@prisma/client";
 import prisma    from "../../../../config/prismaClient";
+import { InvalidParamError } from "../../../../errors/InvalidParamError";
+import { QueryError } from "../../../../errors/QueryError";
 
 class MusicService {
     async create(body: Music) {
+        if (!body) {
+            throw new InvalidParamError("Não é possível adicionar uma música sem nenhuma informação");
+        }
+
         const music = await prisma.music.create({
             data: {
                 name: body.name,
@@ -21,6 +27,10 @@ class MusicService {
             },
         });
 
+        if (!music) {
+            throw new QueryError("Música não existe no sistema!");
+        }
+
         return music;
     }
 
@@ -31,10 +41,27 @@ class MusicService {
             }
         });
 
+        if (!musics) {
+            throw new QueryError("Não há músicas cadastradas no sistema!");
+        }
+
         return musics;
     }
 
     async update(body: Music) {
+        if (!body) {
+            throw new InvalidParamError("Não há informações para atualizar música!");
+        }
+
+        const checkMusic = await prisma.music.findUnique({
+            where: {
+                id: body.id
+            }
+        });
+        if (!checkMusic) {
+            throw new InvalidParamError("Música não existe no sistema");
+        }
+
         const music = await prisma.music.update({
             where: {
                 id: body.id
@@ -46,6 +73,15 @@ class MusicService {
     }
 
     async deleteId(id: number) {
+        const checkMusic = await prisma.music.findUnique({
+            where: {
+                id: id
+            }
+        });
+        if (!checkMusic) {
+            throw new InvalidParamError("Música não existe no sistema");
+        }
+        
         const music = await prisma.music.delete({
             where: {
                 id: id
